code-refactoring: separate review form from the ProductReview component
